+++ examples/supabase/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/example-supabase
 
+## 2.8.5
+
+### Patch Changes
+
+- @stackframe/stack@2.8.5
+
+## 2.8.4
+
+### Patch Changes
+
+- @stackframe/stack@2.8.4
+
 ## 2.8.3
 
 ### Patch Changes