+++ packages/init-stack/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @stackframe/init-stack
 
+## 2.8.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.5
+
+## 2.8.4
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.4
+
 ## 2.8.3
 
 ### Patch Changes