+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/stack-shared
 
+## 2.8.5
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.4
+
+### Patch Changes
+
+- Various changes
+
 ## 2.8.3
 
 ### Patch Changes