+++ packages/template/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @stackframe/stack
 
+## 2.8.5
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.5
+  - @stackframe/stack-ui@2.8.5
+  - @stackframe/stack-sc@2.8.5
+
+## 2.8.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.4
+  - @stackframe/stack-ui@2.8.4
+  - @stackframe/stack-sc@2.8.4
+
 ## 2.8.3
 
 ### Patch Changes